Complete SOT Page to 'Send' & 'Receive' Data from C++ App

+++ sot.js
@@ -13,6 +13,11 @@ module.exports = {
     connect
 }
 
+let listOfData
+var dataCounter = 0
+let prior
+let listOfGroundTruth
+
 function init(win) {
 
     mainWindow = win
@@ -20,31 +25,31 @@ function init(win) {
         
         connect()
 
-        const listOfData = readDataFile('/Users/mohsen/Documents/state_estimate_ws/MOT/stonesoup/PDA/sot_data.txt')
-        const prior = getPrior('/Users/mohsen/Documents/state_estimate_ws/MOT/stonesoup/PDA/sot_data.txt')
-        console.log(prior[0])
-        console.log(prior[1])
-        const listOfGroundTruth = readGroundTruthFile('/Users/mohsen/Documents/state_estimate_ws/MOT/stonesoup/PDA/ground_truth.txt')
-        loopThroughSplittedText(listOfData, listOfGroundTruth)
+        listOfData = readDataFile('/Users/mohsen/Documents/state_estimate_ws/MOT/stonesoup/PDA/sot_data.txt')
+        prior = getPrior('/Users/mohsen/Documents/state_estimate_ws/MOT/stonesoup/PDA/sot_data.txt')
+        
+        listOfGroundTruth = readGroundTruthFile('/Users/mohsen/Documents/state_estimate_ws/MOT/stonesoup/PDA/ground_truth.txt')
+        // loopThroughSplittedText(listOfData, listOfGroundTruth)
 
-        client.write(prior[0] + '|' + prior[1])
+        sendPrior(prior)
+        
 
         clearTimeout(timeoutObj)
 
     }, 2000)
-
 }
+
 
  
   
 function loopThroughSplittedText(values, ground_truth) {
-    for (var i = 0; i < values.length; i++) {
+    // for (var i = 0; i < values.length; i++) {
+    for (var i = 0; i < 2; i++) {
         // for each iteration console.log a word
         // and make a pause after it
         (function (i) {
             setTimeout(function () {
-                // document.getElementById('text').innerHTML += splittedText[i];
-                // console.log(values[i]);
+
                 mainWindow.webContents.send('sot:draw', { 
                     type: 'circle', 
                     x: values[i]['x'], 
@@ -135,13 +140,64 @@ var client = new net.Socket();
 function connect() {
     client.connect(6060, '127.0.0.1', function() {
         console.log('Connected');
-        // client.write('40 | Hello');
     });
 }
 
+
+function sendPrior(prior) {
+    client.write(prior[0] + '|' + prior[1] + '|40:')
+}
+
+function sendMeasurement(measurement, dt) {
+    client.write(measurement + '|24:' + dt + '|89:')
+}
+
 client.on('data', function(data) {
-	console.log('Received: ' + data);
-	client.destroy(); // kill client after server's response
+	
+    if(data == '50') {
+        console.log("State 1 Finished")
+        
+        sendMeasurement(listOfData[dataCounter]['measurements'], listOfData[dataCounter]['dt'])
+    } else {
+
+        data = data.toString()
+        data = data.split('|')
+        var x = data[0].replaceAll('\n', ' ')
+        x = x.split(' ').filter( element => element != '')
+        var covar = data[1].replaceAll('\n', ' ')
+        covar = covar.split(' ').filter( element => element != '')
+        covar = covar.join(', ')
+        
+        mainWindow.webContents.send('sot:draw', { 
+            type: 'circle', 
+            x: x[0],
+            y: x[2],
+            r: 5,
+            covar: covar,
+            measurements: listOfData[dataCounter]['measurements'], 
+            truthX: listOfGroundTruth[dataCounter]['x'],
+            truthY: listOfGroundTruth[dataCounter]['y']
+        })
+
+        
+        dataCounter += 1
+        if(dataCounter < listOfData.length) {
+
+            const timeoutObj = setTimeout(function () {
+        
+                sendMeasurement(listOfData[dataCounter]['measurements'], listOfData[dataCounter]['dt']) 
+                clearTimeout(timeoutObj)
+        
+            }, 1000)
+
+            // sendMeasurement(listOfData[dataCounter]['measurements'], listOfData[dataCounter]['dt'])
+
+        } else {
+            console.log("Operation Finished")
+            client.destroy(); // kill client after server's response
+        }
+    }
+	
 });
 
 client.on('close', function() {
